// src/domain/enums/error-codes.ts
export enum ErrorCodes {
  NOT_FOUND = 'NOT_FOUND_ERROR',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  BAD_REQUEST = 'BAD_REQUEST_ERROR',
  DOMAIN_ERROR = 'DOMAIN_ERROR',
  CONFLICT_ERROR = 'CONFLICT_ERROR',
  DUPLICATE_ENTITY = 'DUPLICATE_ENTITY_ERROR',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY_ERROR',
  INVALID_PARAM = 'INVALID_PARAM_ERROR',
  EXTERNAL_API_ERROR = 'EXTERNAL_API_ERROR'
}